fixed on empty process list

+++ respawn.sh
@@ -1,1 +1,3 @@
-for file in ./process/*; do rm "$file"; bash start.sh "${file##*/}"; done
+if [ "$(ls -A ./process)" ]; then
+	for file in ./process/*; do rm "$file"; bash start.sh "${file##*/}"; done
+fi
